Documents/interface_test/requirements.txt
PyMysql==0.9.2
PyYAML==3.12
requests==2.18.4
xlrd==1.0.0